+++ Hexagon/Hexagon.csproj
@@ -40,12 +40,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="FractionalHex.cs" />
     <Compile Include="Hex.cs" />
     <Compile Include="Layout.cs" />
+    <Compile Include="Maths.cs" />
     <Compile Include="OffsetCoord.cs" />
     <Compile Include="Orientation.cs" />
+    <Compile Include="PointD.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RedBlobLibrary.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
